backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pyodbc>=5.1.0
python-dotenv>=1.0.1
pydantic>=2.7.0
google-genai>=0.6.0
